Moved data accesors to new module `sc.get`
Based on discussion from in: https://github.com/theislab/scanpy/issues/562
